experimento-seguridad/aim/requirements.txt
flask
python-dotenv
pyjwt
Flask-SQLAlchemy